+++ module/ShortcutPanel.js
@@ -1,6 +1,6 @@
-import { ConfigSheetForShortcuts } from "./ConfigSheetForShortcuts.js";
 import { DEFAULT_SHORTCUT_SETTINGS, GlobalConfiguration } from "./constants.js";
-import { CustomCardGUIWrapper } from "./CustomCardGUIWrapper.js";
+import { ConfigSheetForShortcuts } from "./config/ConfigSheetForShortcuts.js";
+import { CustomCardGUIWrapper } from "./mainui/CustomCardGUIWrapper.js";
 
 const HEIGHT_FOR_ONE_CARD = 772;
 const WIDTH_FOR_ONE_CARD = 510;